WIP
- Benchmark class

- Threadpool with coroutine handles

+++ include/result.hpp
@@ -2,12 +2,13 @@
 
 #include <string>
 
-struct Result {
+struct Result
+{
   std::string name;
   size_t threads;
   size_t buffer_size;
   size_t events;
   size_t n;
-  float mean;
-  float std;
+  double mean;
+  double std;
 };
